data-logger/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
dnspython==2.6.1
idna==3.7
pymongo==4.8.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2
paho-mqtt==2.1.0
pymongo==4.8.0